Prepare campaign for C21_WMDE_Test_18
https://phabricator.wikimedia.org/T297206
