+++ desktop/styles/styles.pcss
@@ -10,6 +10,7 @@
 @import './Footer.pcss';
 @import './Infobox.pcss';
 @import './BannerText.pcss';
+@import './AnimateTextHighlight.pcss';
 @import './ProgressBar.pcss';
 @import './Slider.pcss';
 @import '../../shared/components/ui/use_of_funds/FundsModal.pcss';
