--- desktop/styles/styles_var.pcss
@@ -1,19 +1,0 @@
-@import '../../shared/components/ui/use_of_funds/variables.pcss';
-@import 'variables.pcss';
-
-@import '../../shared/components/BannerTransition.pcss';
-
-@import './Banner.pcss';
-@import './DonationForm.pcss';
-@import './DonationFormSteps.pcss';
-@import './SelectGroup.pcss';
-@import './Footer.pcss';
-@import './Infobox.pcss';
-@import './BannerText.pcss';
-@import './AnimateTextHighlight.pcss';
-@import './ProgressBar.pcss';
-@import './Slider.pcss';
-@import '../../shared/components/ui/use_of_funds/FundsModal.pcss';
-@import '../../shared/components/ui/use_of_funds/FundsContent.pcss';
-@import '../../shared/components/ui/use_of_funds/FundsDistributionInfo.pcss';
-@import '../../shared/components/ui/use_of_funds/CompanyBudgets.pcss';
